Move CustomJSONField to a common class and use with the inspection result RawFact
